+++ app/craters/container.js
@@ -1,8 +1,6 @@
 (function(window){ "use strict";
 
 var physics = require ('./physics.js'),
-	update = require ('./update.js'),
-	render = require ('./render.js'),
 	loop = require ('./loop.js'),
 	canvas = require ('./canvas.js');
 	
@@ -14,6 +12,7 @@ var container = class container {
 			    x: 0, 
 			    y: 100
 		    },
+		    
 		    width:  width,
 		    height: height,
 		    frames: frames,
@@ -37,14 +36,46 @@ var container = class container {
 		
 		// Instantiate core modules with the current scope
 		this.physics = new physics( this );
-		this.update = new update( this );
-		this.render = new render( this );
-		this.loop   = new loop( this );
-		this.init();
+		this.loop = new loop( this );
+		this.intitiate ();
 	}
 	
-	init(){
+	intitiate () {
 		
+	}
+	
+	update (scope, now) {
+		
+		var state = scope.state || [];
+		var entities = state.entities;
+		for (var entity = 0; entity < entities.length; entity++) {
+			// Fire off each active entities `render` method
+			entities[entity].update();
+		}
+	}
+	
+	render (scope, now) {
+		// Setup globals
+		var w = scope.constants.width,
+			h = scope.constants.height;
+		
+		// Clear out the canvas
+		scope.context.clearRect(0, 0, w, h);
+		
+		// Spit out some text
+		scope.context.fillStyle = '#ff0';
+		// If we want to show the FPS, then render it in the top right corner.
+		if (scope.constants.debug) {
+			scope.context.fillText('fps : ' + scope.loop.fps, w - 100, 50);
+		}
+		
+		// If there are entities, iterate through them and call their `render` methods
+		var state = scope.state || [];
+		var entities = state.entities;
+		for (var entity = 0; entity < entities.length; entity++) {
+			// Fire off each active entities `render` method
+			entities[entity].render();
+		}
 	}
 };