+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 #### v0.0.0.1
 * miniature physics engine acceleration velocity gravity
+* container has a loop update render cycle if it's own